rename component to be more generic for use in other classes; update references
